Move dtos to dto package
