--- src/main/kotlin/com/hedvig/rapio/externalservices/memberService/CreateMemberResponse.kt
@@ -1,5 +1,0 @@
-package com.hedvig.rapio.externalservices.memberService
-
-data class CreateMemberResponse(
-    val memberId: Long
-)
