+++ src/main/kotlin/com/hedvig/rapio/externalservices/memberService/dto/CreateMemberResponse.kt
@@ -1,0 +1,5 @@
+package com.hedvig.rapio.externalservices.memberService.dto
+
+data class CreateMemberResponse(
+    val memberId: Long
+)
