+++ src/main/kotlin/com/hedvig/rapio/externalservices/memberService/dto/StartOnboardingWithSSNRequest.kt
@@ -1,4 +1,4 @@
-package com.hedvig.rapio.externalservices.memberService
+package com.hedvig.rapio.externalservices.memberService.dto
 
 data class StartOnboardingWithSSNRequest(
     val ssn: String
