+++ src/main/kotlin/com/hedvig/rapio/externalservices/memberService/dto/UpdateContactInformationRequest.kt
@@ -1,4 +1,4 @@
-package com.hedvig.rapio.externalservices.memberService
+package com.hedvig.rapio.externalservices.memberService.dto
 
 import java.time.LocalDate
 
@@ -10,4 +10,12 @@ class UpdateContactInformationRequest(
     val phoneNumber: String? = null,
     val address: Address? = null,
     val birthDate: LocalDate? = null
+)
+
+data class Address(
+    val street: String?,
+    val city: String?,
+    val zipCode: String?,
+    val apartmentNo: String?,
+    val floor: Int?
 )
